Move long_url service to `services/long_url.py`
Moves custom exception to eexceptions, updates tests.
